added bond valuation functionality

+++ main.py
@@ -1,8 +1,11 @@
 from datetime import datetime
 import math
+import os
 from typing import List
 
 from functools import partial
+
+from bs4 import BeautifulSoup
 
 from matplotlib import pyplot as plt
 from matplotlib.animation import FuncAnimation
@@ -93,22 +96,36 @@ def parse_trades(tradedate, market='shares', price_filter=''):
     r = s.post(url, data={'date': tradedate})
     with open('output.html', 'w', encoding='utf8') as f:
         f.write(r.text)
-    df = pd.read_html(r.text)[0]
+    # df = pd.read_html(r.text)[0]
+    df = parse_trades_html(r.text)
     df = df.loc[~(df.iloc[:,1].str.lower().str.contains('итого'))]
     df['tradedate'] = tradedate
-    df.to_csv(f'trades/{market}_{price_filter}_{tradedate}.csv', index=False, sep=';')
+    cols_float = ['Лучшая котировка на покупку', 'Лучшая котировка на продажу',
+       'Цена первой сделки', 'Максимальная цена', 'Минимальная цена',
+       'Цена последней сделки', 'Изменение цены за день, %',
+       'Средневзвеш. цена', 'Объем, млн KZT']
+    #print(df[df.columns[2:-3]])
+    for col in cols_float:
+        df[col] = df[col].str.replace(',','.').str.replace("–", "0").str.replace(" ", "").astype('float64')
+    # print(df.columns)
+    df['Количество сделок'] = df['Количество сделок'].str.replace("–","0").astype('int64')
+    df.to_csv(f'trades/{market}_{price_filter}_{tradedate}.csv', 
+              index=False, sep=';', encoding='cp1251')
     return df
 
 
 def parse_sec_info(isin: str):
-    url = f"https://kase.kz/ru/gsecs/show/{isin}/"
-    s = requests.session()
-    s.headers.update(HEADERS)
-    r = s.get(url)
-    with open(f'sec_info_{isin}.html', 'w', encoding='utf8') as f:
-        f.write(r.text)
-    
-    # print(pd.read_html(r.text)[-1])
+    file_path = f"sec_info/{isin}.csv"
+    if not os.path.exists(file_path):
+        url = f"https://kase.kz/ru/gsecs/show/{isin}/"
+        s = requests.session()
+        s.headers.update(HEADERS)
+        r = s.get(url)
+        with open(f'sec_info_{isin}.html', 'w', encoding='utf8') as f:
+            f.write(r.text)
+        
+        return parse_sec_info_html(r.text)
+    return pd.read_csv(file_path) 
 
 
 
@@ -149,10 +166,57 @@ def create_df_from_params_vect():
     return pd.concat([params['tradedate'], df], axis=1)
 
 
+def parse_trades_html(content: str = None):
+    if not content:
+        print("Content is null. Reading file..")
+        with open('output.html', 'r', encoding='utf8') as f:
+            content = f.read()
+    soup = BeautifulSoup(content, features='lxml')
+    columns =[[col.text.strip() for col in row.find_all('th')] 
+              for row in soup.find('thead').find_all('tr')][0] 
+    data = [{columns[i]:col.text.strip() for i, col in enumerate(row.find_all('td'))} 
+           for row in soup.find('tbody').find_all('tr') 
+           if "Итого" not in row.find('td').text]
+    return pd.DataFrame(data)
+
+
+def parse_sec_info_html(content: str):
+    soup = BeautifulSoup(content, features='lxml')
+    data = [[k.text.strip() for k in row.find_all(class_='info-table__cell')] 
+            for row in soup.find(class_='info-table')\
+            .find_all(class_='info-table__row')]
+    data_dict = {row[0]:row[1] for row in data}
+    coupons = soup.find_all(class_='modal-content')[0]
+    columns =[[col.text.strip() for col in row.find_all('th')] 
+              for row in coupons.find('thead').find_all('tr')][0] 
+    data = [{columns[i]:col.text.strip() 
+             for i, col in enumerate(row.find_all('td'))} 
+           for row in coupons.find('tbody').find_all('tr')]
+    coupons_df = pd.DataFrame(data)
+    coupons_df["Дата начала купонной выплаты"] =  \
+        pd.to_datetime(coupons_df["Дата начала купонной выплаты"].apply(
+        lambda x: f"{x.split('.')[0]}.{x.split('.')[1]}.20{x.split('.')[2]}"), 
+                         format="%d.%m.%Y" )
+    freq = coupons_df["Дата начала купонной выплаты"].dt.year.value_counts().mean()
+    coupons_df["Список ценных бумаг"] = data_dict["Список ценных бумаг:"]
+    coupons_df["Валюта котирования"] = data_dict["Валюта котирования:"]
+    coupons_df["ISIN"] = data_dict["ISIN:"]
+    coupons_df["ISIN (144А)"] = data_dict["ISIN (144А):"]
+    coupons_df["Дата погашения"] = data_dict["Дата погашения:"]
+    coupons_df['freq'] = freq
+    coupons_df['Ставка, % год.'] = coupons_df['Ставка, % год.'].str.replace(',','.').astype('float64')
+    
+    coupons_df.to_csv(f'sec_info/{data_dict["Код бумаги:"]}.csv', 
+                      index=False, 
+                      date_format="%d.%m.%Y")
+    
+    return coupons_df
+    
+
 if __name__=='__main__':
     # download_gcurve_params()
     # animate_gcurve()
-    # parse_trades('16.11.2022', 'kase-global', '')
+    # parse_trades('16.11.2022', 'gsecs', '#gsec_clean')
     parse_sec_info('KZ_06_4410')
     # plot_gcurve(datetime(2022, 11, 16))
     # plot_gcurve_last()
@@ -168,3 +232,4 @@ if __name__=='__main__':
     # print(create_df_from_params_vect().head())
     # toc = time.perf_counter()
     # print(f'func speed is {toc-tic:0.4f} seconds')
+    # parse_trades_html()
